+++ Libraries/MiKTeX/Core/Utils/unx/unxUtils.cpp
@@ -1,6 +1,6 @@
 /* unxUtil.cpp: 
 
-   Copyright (C) 1996-2018 Christian Schenk
+   Copyright (C) 1996-2019 Christian Schenk
 
    This file is part of the MiKTeX Core Library.
 
@@ -136,12 +136,12 @@ bool Utils::CheckPath(bool repair)
     return false;
   }
   
-  PathName localBinDir = session->GetSpecialPath(SpecialPath::LocalBinDirectory);
+  PathName linkTargetDirectory = session->GetSpecialPath(SpecialPath::LinkTargetDirectory);
 
   string repairedPath;
   bool pathCompetition;
   
-  bool pathOkay = !Directory::Exists(localBinDir) || !FixProgramSearchPath(envPath, localBinDir, true, repairedPath, pathCompetition);
+  bool pathOkay = !Directory::Exists(linkTargetDirectory) || !FixProgramSearchPath(envPath, linkTargetDirectory, true, repairedPath, pathCompetition);
 
   bool repaired = false;